+++ gems/aws-sdk-iotfleetwise/lib/aws-sdk-iotfleetwise/client.rb
@@ -3145,7 +3145,7 @@ module Aws::IoTFleetWise
         params: params,
         config: config)
       context[:gem_name] = 'aws-sdk-iotfleetwise'
-      context[:gem_version] = '1.2.0'
+      context[:gem_version] = '1.3.0'
       Seahorse::Client::Request.new(handlers, context)
     end