+++ gems/aws-sdk-medialive/CHANGELOG.md
@@ -1,6 +1,11 @@
 Unreleased Changes
 ------------------
 
+1.93.0 (2022-12-09)
+------------------
+
+* Feature - Link devices now support buffer size (latency) configuration. A higher latency value means a longer delay in transmitting from the device to MediaLive, but improved resiliency. A lower latency value means a shorter delay, but less resiliency.
+
 1.92.0 (2022-12-02)
 ------------------